+++ scripts/packages/internal/index.d.ts
@@ -6,7 +6,10 @@ export {
   HTMLStencilElement,
   JSXBase,
   OutputTargetCustom,
+  OutputTargetDocsCustom,
   Plugin,
   PluginCtx,
   PrintLine
 } from '../dist/declarations';
+export * from '../dist/declarations/docs';
+