+++ src/compiler/entries/resolve-component-dependencies.ts
@@ -1,15 +1,15 @@
 import * as d from '../../declarations';
 import { flatOne, unique } from '@utils';
 
-export function resolveComponentDependencies(cmps: d.ComponentCompilerMeta[]) {
-  computeDependencies(cmps);
+export function resolveComponentDependencies(compilerCtx: d.CompilerCtx, cmps: d.ComponentCompilerMeta[]) {
+  computeDependencies(compilerCtx, cmps);
   computeDependants(cmps);
 }
 
-function computeDependencies(cmps: d.ComponentCompilerMeta[]) {
+function computeDependencies(compilerCtx: d.CompilerCtx, cmps: d.ComponentCompilerMeta[]) {
   const visited = new Set();
   cmps.forEach(cmp => {
-    resolveTransitiveDependencies(cmp, cmps, visited);
+    resolveTransitiveDependencies(compilerCtx, cmp, cmps, visited);
     cmp.dependencies = unique(cmp.dependencies).sort();
   });
 }
@@ -20,19 +20,21 @@ function computeDependants(cmps: d.ComponentCompilerMeta[]) {
   });
 }
 
-function resolveTransitiveDependencies(cmp: d.ComponentCompilerMeta, cmps: d.ComponentCompilerMeta[], visited: Set<d.ComponentCompilerMeta>): string[] {
+function resolveTransitiveDependencies(compilerCtx: d.CompilerCtx, cmp: d.ComponentCompilerMeta, cmps: d.ComponentCompilerMeta[], visited: Set<d.ComponentCompilerMeta>): string[] {
   if (visited.has(cmp)) {
     return cmp.dependencies;
   }
   visited.add(cmp);
 
-  const dependencies = cmp.potentialCmpRefs.filter(tagName => cmps.some(c => c.tagName === tagName));
+  const moduleFile = compilerCtx.moduleMap.get(cmp.sourceFilePath);
+
+  const dependencies = moduleFile.potentialCmpRefs.filter(tagName => cmps.some(c => c.tagName === tagName));
   cmp.dependencies = cmp.directDependencies = dependencies;
 
   const transitiveDeps = flatOne(
     dependencies
       .map(tagName => cmps.find(c => c.tagName === tagName))
-      .map(c => resolveTransitiveDependencies(c, cmps, visited))
+      .map(c => resolveTransitiveDependencies(compilerCtx, c, cmps, visited))
   );
   return cmp.dependencies = [
     ...dependencies,