+++ src/compiler/transformers/static-to-meta/visitor.ts
@@ -1,8 +1,10 @@
 import * as d from '../../../declarations';
 import { getModule, resetModule } from '../../build/compiler-ctx';
-import ts from 'typescript';
-import { parseStaticComponentMeta } from './component';
+import { parseCallExpression } from './call-expression';
 import { parseImport } from './import';
+import { parseStaticComponentMeta } from './component';
+import { parseStringLiteral } from './string-literal';
+import ts from 'typescript';
 
 
 export function convertStaticToMeta(config: d.Config, compilerCtx: d.CompilerCtx, buildCtx: d.BuildCtx, typeChecker: ts.TypeChecker, collection: d.CollectionCompilerMeta, transformOpts: d.TransformOptions): ts.TransformerFactory<ts.SourceFile> {
@@ -14,9 +16,13 @@ export function convertStaticToMeta(config: d.Config, compilerCtx: d.CompilerCtx
 
     function visitNode(node: ts.Node): ts.VisitResult<ts.Node> {
       if (ts.isClassDeclaration(node)) {
-        return parseStaticComponentMeta(config, transformCtx, typeChecker, node, moduleFile, compilerCtx.nodeMap, transformOpts);
+        return parseStaticComponentMeta(config, compilerCtx, transformCtx, typeChecker, node, moduleFile, compilerCtx.nodeMap, transformOpts);
       } else if (ts.isImportDeclaration(node)) {
         return parseImport(config, compilerCtx, buildCtx, moduleFile, dirPath, node);
+      } else if (ts.isCallExpression(node)) {
+        parseCallExpression(moduleFile, node);
+      } else if (ts.isStringLiteral(node)) {
+        parseStringLiteral(moduleFile, node);
       }
       return ts.visitEachChild(node, visitNode, transformCtx);
     }