+++ src/dev-server/dev-client/test/build-status.spec.ts
@@ -1,4 +1,4 @@
-import { ICON_ERROR, ICON_TYPE, initFavIcons, updateFavIcon } from '../build-status';
+import { ICON_ERROR, ICON_TYPE, initBuildStatus, updateFavIcon } from '../build-status';
 
 
 describe('build-status', () => {
@@ -10,7 +10,7 @@ describe('build-status', () => {
     linkElm.rel = 'shortcut icon';
     linkElm.type = 'org-type';
     document.head.appendChild(linkElm);
-    initFavIcons(document);
+    initBuildStatus(window as any, document);
 
     expect(linkElm.dataset.href).toBe('http://testing.stenciljs.com/org-icon.ico');
     expect(linkElm.dataset.type).toBe('org-type');